Empty array response produce empty array object
If the program asks about a collection, it should get a collection even
being empty.
In case of an empty hash, implies an empty object, so makes sense to
still return nil for this case.
In case of nil as response from the Query object, does not make sense to
treat this as an expected case, so it will be treated as unexpected
response.